influenza-adapter/requirements.txt
pytesseract==0.3.2
redis==3.4.1
requests==2.22.0
beautifulsoup4==4.8.2
Pillow==7.0.0
